Adding OrderDetails component for Order page. Forwards the customer after checking out from shoppingcart to an order page.

+++ backend/src/main/java/at/wst/online_webshop/dtos/CartItemDTO.java
@@ -21,7 +21,7 @@ public class CartItemDTO {
     private Long productId;
     private String productName;
     private String productImageUrl;
-    private Long productPrice;
+    private Double productPrice;
     private String productCategory;
     private int cartItemQuantity;
     private BigDecimal cartItemSubprice;
